requirements.txt
appdirs==1.4.4
appnope==0.1.2
asgiref==3.4.1
attrs==21.2.0
backcall==0.2.0
beautifulsoup4==4.10.0
black==21.7b0
boto3==1.21.28
botocore==1.24.28
Brotli==1.0.9
certifi==2021.5.30
charset-normalizer==2.0.3
click==8.0.1
colorama==0.4.4
coverage==5.5
cssselect2==0.4.1
dash==2.0.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-renderer==1.9.1
dash-table==5.0.0
decorator==5.0.9
Django==4.0.2
django-autoslug==1.9.8
django-cors-headers==3.10.0
django-countries==7.2.1
django-dash==0.6.1
django-extensions==3.1.3
django-filter==2.4.0
django-nine==0.2.5
django-ses==2.6.1
django-storages==1.12.3
django-tinymce==3.4.0
djangorestframework==3.12.4
dnspython==2.2.0
docopt==0.6.2
docutils==0.17.1
dpd-components==0.1.0
dropbox==11.27.0
etsy2==0.7.0
fastapi==0.66.1
feedparser==6.0.8
fett==0.3.2
flake8==4.0.1
Flask==2.0.2
Flask-Compress==1.10.1
freezegun==1.1.0
future==0.18.2
gunicorn==20.1.0
h11==0.12.0
idna==3.2
iniconfig==1.1.1
ipdb==0.13.9
ipython==7.25.0
ipython-genutils==0.2.0
isort==5.9.3
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.3
jmespath==0.10.0
lazy-object-proxy==1.7.1
lxml==4.6.4
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mccabe==0.6.1
mypy-extensions==0.4.3
networkx==2.6.3
numpy==1.21.4
oauthlib==3.1.1
packaging==21.0
pandas==1.3.4
parso==0.8.2
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
pif==0.8.2
Pillow==8.4.0
platformdirs==2.4.1
plotly==5.4.0
pluggy==0.13.1
ply==3.11
prompt-toolkit==3.0.19
psycopg2==2.9.1
ptyprocess==0.7.0
py==1.10.0
pycodestyle==2.8.0
pydantic==1.8.2
pyflakes==2.4.0
Pygments==2.9.0
pymongo==3.12.3
pyparsing==2.4.7
pytest==6.2.5
pytest-cov==3.0.0
pytest-django==4.5.2
python-dateutil==2.8.2
python-jsonrpc-server==0.3.4
pytz==2021.1
PyYAML==5.4.1
regex==2021.7.6
reportlab==3.6.3
requests==2.26.0
requests-oauthlib==1.3.0
s3transfer==0.5.0
sgmllib3k==1.0.0
six==1.16.0
snooty-lextudio==1.11.6
soupsieve==2.2.1
sqlparse==0.4.2
starlette==0.14.2
stone==3.3.1
svglib==1.1.0
tenacity==8.0.1
tinycss2==1.1.1
toml==0.10.2
tomli==1.0.4
traitlets==5.0.5
transliterate==1.10.2
typing-extensions==3.10.0.0
ujson==1.35
Unidecode==1.3.2
urllib3==1.26.6
uvicorn==0.14.0
vishap==0.1.5
watchdog==1.0.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
